+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import '../src/assets/styles/reset.css';
 import '../src/assets/styles/index.css';
 import '../src/assets/styles/color.css';
-import IndexPage from './components/indexPage';
+import IndexPage from './components/index/indexPage';
 
 function App() {
   return (
