+++ src/services/folderService.js
@@ -1,4 +1,4 @@
-import api from './api';
+import api from '../utils/api';
 
 export const getFolderData = async () => {
   const response = await api.get('/sample/folder');
